Fix freeze on upload issue
For some reason, the board was freezing when uploading code. The program
memory was 31100 bytes, and either removing or adding a print statement
fixed the problem. I have no explanation for why this would happen at
this time. We can come back to this commit to recreate the issue and
try to find the real source of the problem.